+++ NEWS.md
@@ -7,6 +7,13 @@ output: html_document
 Cite all releases with doi [10.5281/zenodo.5226085](https://doi.org/10.5281/zenodo.5226085), 
 which always resolves to the latest release.
 
+* Testing now done in parallel.
+* New function `extend_fu_details_to_exergy()` converts
+  the final-to-useful details matrices from energy to exergy.
+* New tests for new features.
+    - Now up to 759 tests, all passing.
+    - Test coverage remains at 100%.
+
 
 # Recca 0.1.54 (2023-12-21) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10420370.svg)](https://doi.org/10.5281/zenodo.10420370)